requirements.txt
attrs==19.1.0
backcall==0.1.0
bleach==3.1.0
bokeh==1.1.0
Click==7.0
cloudpickle==0.8.1
colorama==0.4.1
cycler==0.10.0
dask==1.2.1
decorator==4.4.0
defusedxml==0.6.0
distributed==1.27.1
entrypoints==0.3
extras==1.0.0
fixtures==3.0.0
graphviz==0.10.1
HeapDict==1.0.0
ipykernel==5.1.0
ipython==7.5.0
ipython-genutils==0.2.0
ipywidgets==7.4.2
jedi==0.13.3
Jinja2==2.10.1
joblib==0.13.2
jsonschema==3.0.1
jupyter==1.0.0
jupyter-client==5.2.4
jupyter-console==6.0.0
jupyter-core==4.4.0
jupyterlab==0.35.6
jupyterlab-server==0.2.0
kiwisolver==1.1.0
linecache2==1.0.0
locket==0.2.0
MarkupSafe==1.1.1
matplotlib==3.0.3
mistune==0.8.4
msgpack==0.6.1
nbconvert==5.5.0
nbformat==4.4.0
notebook==5.7.8
numpy==1.16.3
packaging==19.0
pandas==0.24.2
pandocfilters==1.4.2
parso==0.4.0
partd==0.3.10
pbr==5.2.0
pexpect==4.7.0
pickleshare==0.7.5
Pillow==6.0.0
prometheus-client==0.6.0
prompt-toolkit==2.0.9
psutil==5.6.2
ptyprocess==0.6.0
Pygments==2.3.1
pyparsing==2.4.0
pyrsistent==0.15.1
python-dateutil==2.8.0
python-mimeparse==1.6.0
pytz==2019.1
pywinpty==0.5.5
PyYAML==5.1
pyzmq==18.0.1
qtconsole==4.4.3
scikit-learn==0.20.3
scipy==1.2.1
Send2Trash==1.5.0
six==1.12.0
sklearn==0.0
sortedcontainers==2.1.0
streamz==0.5.0
tblib==1.3.2
terminado==0.8.2
testpath==0.4.2
testtools==2.3.0
toolz==0.9.0
tornado==6.0.2
traceback2==1.4.0
traitlets==4.3.2
unittest2==1.1.0
wcwidth==0.1.7
webencodings==0.5.1
widgetsnbextension==3.4.2
zict==0.1.4
